Deprecate Hydra::ModelMethods extract apply_depositor_metadata to Hydra::WithDepositor
