+++ hydra-access-controls/spec/spec_helper.rb
@@ -28,6 +28,7 @@ require_relative '../app/models/hydra/access_controls/access_control_list'
 require_relative '../app/models/hydra/access_controls/permission'
 require_relative '../app/models/hydra/access_controls/embargo'
 require_relative '../app/models/hydra/access_controls/lease'
+require_relative '../app/models/concerns/hydra/with_depositor'
 require_relative '../app/services/hydra/lease_service'
 require_relative '../app/services/hydra/embargo_service'
 require_relative '../app/validators/hydra/future_date_validator'
